Refactor: 모달 제목 , 모달 열기 버튼 리펙터링(#7)
