+++ src/components/modalForm/CategoryForm.tsx
@@ -6,7 +6,7 @@ import Button from '@mui/material/Button';
 import { EDITCATEGORY } from 'src/assets/data/modal/modals';
 import { TextField } from '@mui/material';
 
-export default function CategoryForm({ modalType }) {
+export default function CategoryForm({ modalType, beforeData }) {
   const CategorySchema = Yup.object().shape({
     [CATEGORY]: Yup.string().required('필수입니다.'),
     [DESCRIPTION]: Yup.string(),
